+++ R/create_episode.R
@@ -1,20 +1,26 @@
 #' Create an Episode from a template
 #'
+#' These functions allow you to create an episode that will be added to the
+#' schedule.
+#'
+#'
 #' @param title the title of the episode
 #' @param ext a character. If `ext = "Rmd"` (default), then the new episode will
 #'   be an R Markdown episode. If `ext = "md"`, then the new episode will be
 #'   a markdown episode, which can not generate dynamic content.
-#' @param make_prefix a logical. If `TRUE`, the prefix for the file will be
-#'   automatically determined by the files already present. Otherwise, it assumes
-#'   you have added the prefix.
+#' @param make_prefix a logical. When `TRUE`, the prefix for the file will be
+#'   automatically determined by the files already present. When `FALSE`
+#'   (default), it assumes no prefix is needed.
 #' @param path the path to the {sandpaper} lesson.
-#' @param add if `TRUE`, the lesson is added to the schedule. Defaults to `FALSE`
+#' @param add (logical or numeric) If numeric, it represents the position the
+#'   episode should be added. If `TRUE`, the episode is added to the end of the 
+#'   schedule. If `FALSE`, the episode is added as a draft episode.
 #' @export
 #' @examples
 #' tmp <- tempfile()
 #' create_lesson(tmp)
 #' create_episode_md("getting-started", path = tmp)
-create_episode <- function(title, ext = "Rmd", make_prefix = TRUE, add = FALSE, path = ".") {
+create_episode <- function(title, ext = "Rmd", make_prefix = FALSE, add = TRUE, path = ".") {
   check_lesson(path)
   ext <- switch(match.arg(tolower(ext), c("rmd", "md")), rmd = ".Rmd", md = ".md")
   prefix <- ""
@@ -29,20 +35,32 @@ create_episode <- function(title, ext = "Rmd", make_prefix = TRUE, add = FALSE, 
   copy_template("episode", fs::path(path, "episodes"), ename, 
     values = list(title = siQuote(title), md = ext == ".md"))
   if (add) {
-    suppressWarnings(sched <- get_episodes(path))
-    set_episodes(path, c(sched, ename), write = TRUE)
+    move_episode(ename, position = add, write = TRUE, path = path)
   }
   invisible(fs::path(path, "episodes", ename))
 }
 
+
 #' @export
 #' @rdname create_episode
-create_episode_md <- function(title, make_prefix = TRUE, add = FALSE, path = ".") {
+create_episode_md <- function(title, make_prefix = FALSE, add = TRUE, path = ".") {
   create_episode(title, ext = "md", make_prefix = make_prefix, add = add, path = path)
 }
 
 #' @export
 #' @rdname create_episode
-create_episode_rmd <- function(title, make_prefix = TRUE, add = FALSE, path = ".") {
+create_episode_rmd <- function(title, make_prefix = FALSE, add = TRUE, path = ".") {
   create_episode(title, ext = "Rmd", make_prefix = make_prefix, add = add, path = path)
+}
+
+#' @export
+#' @rdname create_episode
+draft_episode_md <- function(title, make_prefix = FALSE, path = ".") {
+  create_episode(title, ext = "md", make_prefix = make_prefix, add = FALSE, path = path)
+}
+
+#' @export
+#' @rdname create_episode
+draft_episode_rmd <- function(title, make_prefix = FALSE, path = ".") {
+  create_episode(title, ext = "Rmd", make_prefix = make_prefix, add = FALSE, path = path)
 }