+++ R/utils-renv.R
@@ -127,6 +127,8 @@ renv_setup_profile <- function(path = ".", profile = "lesson-requirements") {
     wd <- getwd()
     on.exit(setwd(wd))
     setwd(path)
+    # NOTE: I do not know why, but this takes forever to run when no internet is
+    # available. Kevin may know why. 
     renv::init(project = path, bare = TRUE, restart = FALSE, profile = profile)
     renv::deactivate(project = path)
   },