+++ tests/testthat/test-set_dropdown.R
@@ -53,10 +53,10 @@ test_that("schedule is empty by default", {
 
   cfg <- get_config(tmp)
   suppressMessages(s <- get_episodes(tmp))
-  expect_equal(s, "01-introduction.Rmd", ignore_attr = TRUE)
+  expect_equal(s, "introduction.Rmd", ignore_attr = TRUE)
   expect_null(set_episodes(tmp, s, write = TRUE))
   expect_silent(s <- get_episodes(tmp))
-  expect_equal(s, "01-introduction.Rmd", ignore_attr = TRUE)
+  expect_equal(s, "introduction.Rmd", ignore_attr = TRUE)
 
   # the config files should be unchanged from the schedule
   no_episodes <- names(cfg)[names(cfg) != "episodes"]
@@ -65,11 +65,11 @@ test_that("schedule is empty by default", {
 })
 
 
-test_that("new episodes will not add to the schedule by default", {
+test_that("new episodes will add to the schedule by default", {
 
-  set_episodes(tmp, "01-introduction.Rmd", write = TRUE)
+  set_episodes(tmp, "introduction.Rmd", write = TRUE)
   create_episode("new", path = tmp)
-  expect_equal(get_episodes(tmp), "01-introduction.Rmd", ignore_attr = TRUE)
+  expect_equal(get_episodes(tmp), c("introduction.Rmd", "new.Rmd"), ignore_attr = TRUE)
 
 })
 
@@ -78,7 +78,7 @@ test_that("get_episodes() returns episodes in dir if schedule is not set", {
 
   reset_episodes(tmp)
   suppressMessages(expect_message(s <- get_episodes(tmp)))
-  expect_equal(s, c("01-introduction.Rmd", "02-new.Rmd"), ignore_attr = TRUE)
+  expect_equal(s, c("introduction.Rmd", "new.Rmd"), ignore_attr = TRUE)
   set_episodes(tmp, s[1], write = TRUE)
   expect_equal(get_episodes(tmp), s[1], ignore_attr = TRUE)
 
@@ -91,7 +91,7 @@ cli::test_that_cli("set_episodes() will display the modifications if write is no
   reset_episodes(tmp)
   expect_snapshot(s <- get_episodes(tmp))
 
-  expect_equal(s, c("01-introduction.Rmd", "02-new.Rmd"))
+  expect_equal(s, c("introduction.Rmd", "new.Rmd"))
   set_episodes(tmp, s, write = TRUE)
   expect_equal(get_episodes(tmp), s, ignore_attr = TRUE)
 
@@ -111,49 +111,49 @@ test_that("set_episodes() will error if no proposal is defined", {
 
 test_that("adding episodes will concatenate the schedule", {
 
-  set_episodes(tmp, "01-introduction.Rmd", write = TRUE)
-  expect_equal(get_episodes(tmp), "01-introduction.Rmd")
+  set_episodes(tmp, "introduction.Rmd", write = TRUE)
+  expect_equal(get_episodes(tmp), "introduction.Rmd")
   create_episode("second-episode", add = TRUE, path = tmp)
   expect_equal(res, tmp, ignore_attr = TRUE)
-  expect_equal(get_episodes(tmp), c("01-introduction.Rmd", "03-second-episode.Rmd"), ignore_attr = TRUE)
+  expect_equal(get_episodes(tmp), c("introduction.Rmd", "second-episode.Rmd"), ignore_attr = TRUE)
 
   sb <- create_sidebar(get_episodes(tmp, trim = FALSE))
   expect_length(sb, 2)
-  expect_match(sb[1], "01-introduction.html")
-  expect_match(sb[2], "03-second-episode.html")
+  expect_match(sb[1], "introduction.html")
+  expect_match(sb[2], "second-episode.html")
 
 })
 
 test_that("the schedule can be rearranged", {
 
-  set_episodes(tmp, c("03-second-episode.Rmd", "01-introduction.Rmd"), write = TRUE)
-  expect_equal(get_episodes(tmp), c("03-second-episode.Rmd", "01-introduction.Rmd"), ignore_attr = TRUE)
+  set_episodes(tmp, c("second-episode.Rmd", "introduction.Rmd"), write = TRUE)
+  expect_equal(get_episodes(tmp), c("second-episode.Rmd", "introduction.Rmd"), ignore_attr = TRUE)
 
   sb <- create_sidebar(get_episodes(tmp, trim = FALSE))
   expect_length(sb, 2)
-  expect_match(sb[1], "03-second-episode.html")
-  expect_match(sb[2], "01-introduction.html")
+  expect_match(sb[1], "second-episode.html")
+  expect_match(sb[2], "introduction.html")
 
 })
 
 test_that("yaml lists are preserved with other schedule updates", {
   
-  set_episodes(tmp, c("03-second-episode.Rmd", "01-introduction.Rmd"), write = TRUE)
+  set_episodes(tmp, c("second-episode.Rmd", "introduction.Rmd"), write = TRUE)
   # regression test for https://github.com/carpentries/sandpaper/issues/53
-  expect_equal(get_episodes(tmp), c("03-second-episode.Rmd", "01-introduction.Rmd"))
-  set_learners(tmp, order = "setup.md", write = TRUE) # ZNK 2021-09-07: force this line to be recognized by Git
-  expect_equal(get_episodes(tmp), c("03-second-episode.Rmd", "01-introduction.Rmd"))
+  expect_equal(get_episodes(tmp), c("second-episode.Rmd", "introduction.Rmd"))
+  set_learners(tmp, order = "setup.md", write = TRUE) # ZNK 2021-07: force this line to be recognized by Git
+  expect_equal(get_episodes(tmp), c("second-episode.Rmd", "introduction.Rmd"))
 
 })
 
 test_that("the schedule can be truncated", {
 
-  set_episodes(tmp, "01-introduction.Rmd", write = TRUE)
-  expect_equal(get_episodes(tmp), "01-introduction.Rmd", ignore_attr = TRUE)
+  set_episodes(tmp, "introduction.Rmd", write = TRUE)
+  expect_equal(get_episodes(tmp), "introduction.Rmd", ignore_attr = TRUE)
 
   sb <- create_sidebar(get_episodes(tmp, trim = FALSE))
   expect_length(sb, 1)
-  expect_match(sb[1], "01-introduction.html")
+  expect_match(sb[1], "introduction.html")
 
   skip_if_not(rmarkdown::pandoc_available("2.11"))
 
@@ -166,6 +166,6 @@ test_that("the schedule can be truncated", {
     ".//div[contains(@class, 'accordion-header')]/a")
   links <- xml2::xml_attr(episodes, "href")
   expect_length(links, 1)
-  expect_equal(links[1], "01-introduction.html")
+  expect_equal(links[1], "introduction.html")
 
 })